+++ src/main/java/de/unisb/cs/st/evosuite/coverage/mutation/MutationTimeoutStoppingCondition.java
@@ -33,7 +33,7 @@ public class MutationTimeoutStoppingCondition extends StoppingCondition {
 	 * @see de.unisb.cs.st.evosuite.ga.StoppingCondition#getCurrentValue()
 	 */
 	@Override
-	public int getCurrentValue() {
+	public long getCurrentValue() {
 		return timeouts;
 	}
 
@@ -63,12 +63,12 @@ public class MutationTimeoutStoppingCondition extends StoppingCondition {
 	 * @see de.unisb.cs.st.evosuite.ga.StoppingCondition#setLimit(int)
 	 */
 	@Override
-	public void setLimit(int limit) {
-		MAX_TIMEOUTS = limit;
+	public void setLimit(long limit) {
+		MAX_TIMEOUTS = (int) limit;
 	}
 
 	@Override
-	public int getLimit() {
+	public long getLimit() {
 		return MAX_TIMEOUTS;
 	}
